Improve module augmentation for `word-count` plugin
