+++ packages/ckeditor5-word-count/src/index.ts
@@ -8,3 +8,6 @@
  */
 
 export { default as WordCount, type WordCountUpdateEvent } from './wordcount';
+export { WordCountConfig } from './wordcountconfig';
+
+import './augmentation';
